Cherry picking PR #485, #486 and #487 changes into `support/3.x` (#488)
* MMP-566: Upgrade maven-compiler-plugin of Mule Lifecycle to 3.8.1 (#485)

This version (the latest at the time of writing this) was released on
May 2019 and seems to be pretty stable after 2 years and no new
releases.

* Set maven-jar-plugin version (#487)

Setting the version prevents showing several warnings like the following
one at the start of the building and makes it more stable and
predictable.

```
[WARNING]
[WARNING] Some problems were encountered while building the effective model for org.mule.tools.maven:mule-artifact-tools:pom:3.5.2-SNAPSHOT
[WARNING] 'build.plugins.plugin.version' for org.apache.maven.plugins:maven-jar-plugin is missing. @ line 467, column 21
[WARNING]
[WARNING] It is highly recommended to fix these problems because they threaten the stability of your build.
[WARNING]
[WARNING] For this reason, future Maven versions might no longer support building such malformed projects.
[WARNING]
```

* MMP-567: Fix wrong maven-clean-plugin version for Mule Lifecycle (#486)

The previous version (3.6.1) does not currently exists. At the time of
writing this the latest version is 3.1.0 on Maven Central.