Cherrypick SE-20845, SE-20302, MMP-571, SE-17938 (#497)
* SE-17938: App packaged with maven profile doesn't run.Cannot load class (#490)

* Se-17938: App packaged with maven profile doesn't run.Cannot load class

There was no support for maven profiles. The support is added in runtime
and munit already. This commit complements the fix adding support for
profiles in the mule packager

* test added

* MMP-571: Add support for secure Properties (#492)

* MMP-571: Add support for secure Properties

first commit

* add properties to redeploying

* format (#493)

* SE-20302:CloudHub Deployment Fails with Error 400 Bad Request (#495)

* SE-20302:CloudHub Deployment Fails with Error 400 Bad Request

Removing suffix for deployments because Ch always uses the last patch
for the version (if a suffix is specified it fails). For redeploys we do
nothing because you can deploy to the patch that is already running.

* removing replace to test functionality

* update tests

* SE-20845: Customer unable to deploy to RTF via Maven without assigning user "Manage RTF" permission (#496)

* SE-20845:Customer unable to deploy to RTF via Maven without...

we were using an old API that has a problem with the permissions. We
update it follow the suggestion of rtf team.

* update tests