MMP-570: Generate and include a serializedAST in the packaged app
In order to properly generate the AST, the extension models must be
obtained for the extensions. This in turn requires to create a class
loader for each extension. An approach like this is already implemented
in Tooling and in the extensions packager.

Ideally, if there is an AST already generated (present somewhere in the
target folder) and any of the inputs of the AST were not modified since
that last generation, regenerating the AST should be avoided. Some
conditions under which the AST MUST be regenerated (not a full list)
are: modified any app xml configs, modified the packager version in the
pom, modified a major version of a mule-plugin dependency.