MMP-570:Generate and include a serializedAST in the packaged app (#509)
* MMP-570: Generate and include a serializedAST in the packaged app

In order to properly generate the AST, the extension models must be
obtained for the extensions. This in turn requires to create a class
loader for each extension. An approach like this is already implemented
in Tooling and in the extensions packager.

Ideally, if there is an AST already generated (present somewhere in the
target folder) and any of the inputs of the AST were not modified since
that last generation, regenerating the AST should be avoided. Some
conditions under which the AST MUST be regenerated (not a full list)
are: modified any app xml configs, modified the packager version in the
pom, modified a major version of a mule-plugin dependency.

* MMP-578:Ast generation fails with Cannot load class

and other fixes

* fix packager tests

mule-plugin-a, b y c were empty and the ast generation was failing.

* Resolve transitive mule-plugin ext model

and other small fixes

* MMP-587,MMP-585: Update runtime to 4.5 and add skip property

also:
*tests are fixed
*ast generation reformated to use contentResolver
*artifact.ast moved to mule-artifact folder

* Fix deploy test

* move serialization to mule-extension-model-loader

* fix compile-mojo test and domain problem

* Removing unnecesary classes and other minor changes

* pom fixes

Co-authored-by: pablocabrera85 <pablo.cabrera@mulesoft.com>