MMP-586: Package EE dependencies when building MMP
We need EE dependencies to obtain the ext-models of the mule plugins for
the AST. But the MMP should not need the EE repos configured to be used.
So, these new dependencies are packaged in a
mule-extension-model-loader-dependencies.jar. When mmp is used this jar
is uncompressed and the inner jars are added to the classpath