W-11286314:AST generation fails when having deps with provided scope
we should move the AST generation to de process classes phase.
Generating it after compile phase will allow us to have all the
dependencies resolved (including transitive ones and those with
"provided" or "runtime" scope). Also if the app is already compiled we
can avoid the temporary app for obtaining the extension models.