W-11369839:Mule Maven Plugin AST Generation failing
when certain exception types selected in Error Handler.

Root cause:Mule apikit classes were present in the classloader when
http-connector extension model was resolved. In runtime, each connector
has its own classloader.

Solution: add the jars of the plugins to the classloader (this is needed to get the files to validate AST) AFTER the resolution of all the extension models